+++ src/algorithms/Abml.py
@@ -2,15 +2,20 @@ import torch
 import higher
 import typing
 
-from MLBaseClass import MLBaseClass
-from HyperNetClasses import NormalVariationalNet
-from Maml import Maml
+from src._utils import kl_divergence_gaussians
+from src.HyperNetClasses import NormalVariationalNet
+from src.algorithms.MLBaseClass import MLBaseClass
+from src.algorithms.Maml import Maml
 
-class Vampire2(MLBaseClass):
+class Abml(MLBaseClass):
     def __init__(self, config: dict) -> None:
         super().__init__(config=config)
 
         self.hyper_net_class = NormalVariationalNet
+
+        # prior parameters
+        self.gamma_prior = torch.distributions.gamma.Gamma(concentration=1, rate=0.01)
+        self.normal_prior = torch.distributions.normal.Normal(loc=0, scale=1)
 
     def load_model(self, resume_epoch: int, eps_dataloader: torch.utils.data.DataLoader, **kwargs) -> dict:
         maml_temp = Maml(config=self.config)
@@ -24,13 +29,15 @@ class Vampire2(MLBaseClass):
             track_higher_grads=self.config["train_flag"]
         )
 
+        p_params = [p for p in model["hyper_net"].parameters()]
+
         for _ in range(self.config['num_inner_updates']):
             q_params = f_hyper_net.fast_params # parameters of the task-specific hyper_net
 
             grads_accum = [0] * len(q_params) # accumulate gradients of Monte Carlo sampling
 
             # KL divergence
-            KL_loss = self.KL_divergence_standard_normal(p=q_params)
+            KL_loss = kl_divergence_gaussians(p=q_params, q=p_params)
 
             for _ in range(self.config['num_models']):
                 # generate parameter from task-specific hypernet
@@ -92,9 +99,27 @@ class Vampire2(MLBaseClass):
         
         loss = loss / len(logits)
 
-        # KL loss
-        KL_loss = self.KL_divergence_standard_normal(p=adapted_hyper_net.fast_params)
-        loss = loss + self.config["KL_weight"] * KL_loss
+        # KL divergence
+        KL_div = kl_divergence_gaussians(p=adapted_hyper_net.fast_params, q=[param for param in model["hyper_net"].parameters()])
+
+        loss = loss + self.config["KL_weight"] * KL_div
+
+        # loss prior
+        loss_prior = 0
+        p_params = [param for param in model["hyper_net"].parameters()]
+        for i, param in enumerate(p_params):
+            if i < (len(p_params) // 2):
+                loss_prior = loss_prior - self.normal_prior.log_prob(value=param).sum()
+            else:
+                tau = torch.exp(-2 * param)
+                loss_prior = loss_prior - self.gamma_prior.log_prob(value=tau).sum()
+
+        # regularization is weighted by inverse of the number of mini-batches used.
+        # However, the number of mini-batches might change since one might want to train less or more.
+        # For simplicity, the KL_weight is used as the weighting factor.
+        loss_prior = loss_prior * self.config['KL_weight']
+
+        loss = loss + loss_prior / self.config["minibatch"]
 
         return loss
 
@@ -104,7 +129,12 @@ class Vampire2(MLBaseClass):
 
         logits = self.prediction(x=x_v, adapted_hyper_net=adapted_hyper_net, model=model)
 
-        loss = self.validation_loss(x=x_v, y=y_v, adapted_hyper_net=adapted_hyper_net, model=model)
+        # classification loss
+        loss = 0
+        for logits_ in logits:
+            loss = loss + self.config['loss_function'](input=logits_, target=y_v)
+        
+        loss = loss / len(logits)
 
         y_pred = 0
         for logits_ in logits:
@@ -115,24 +145,3 @@ class Vampire2(MLBaseClass):
         accuracy = (y_pred.argmax(dim=1) == y_v).float().mean().item()
 
         return loss.item(), accuracy * 100
-
-    @staticmethod
-    def KL_divergence_standard_normal(p: typing.List[torch.Tensor]) -> typing.Union[torch.Tensor, float]:
-        """Calculate KL divergence between a diagonal Gaussian with N(0, I)
-        """
-        KL_div = 0
-
-        n = len(p) // 2
-
-        for i in range(n):
-            p_mean = p[i]
-            p_log_std = p[n + i]
-
-            KL_div = KL_div + torch.sum(input=torch.square(input=p_mean))
-            KL_div = KL_div + torch.sum(input=torch.exp(input=2 * p_log_std))
-            KL_div = KL_div - n
-            KL_div = KL_div - 2 * torch.sum(input=p_log_std)
-
-        KL_div = KL_div / 2
-
-        return KL_div
